Refactored c mongo download in bash and Node.js.
